Move random name generator to common place
